Changed MITNewsRefreshControlHangTime to float. Added 5 minute refresh to category view.

+++ Modules/News/View Controllers/MITNewsConstants.h
@@ -23,7 +23,7 @@ extern NSString* const MITNewsStoryExternalNoImageCellNibName;
 
 extern NSString* const MITNewsCategoryHeaderIdentifier;
 extern NSUInteger const MITNewsDefaultNumberOfStoriesPerPage;
-extern NSUInteger const MITNewsRefreshControlHangTime;
+extern CGFloat const MITNewsRefreshControlHangTime;
 
 #pragma mark iPad stories
